Optimize Loads bypassing trivial unrelated array stores.
Fix some graphviz display on array loads & stores
Fix bug in GCM, extra code for infinite loops now handled elsewhere.
Move BrainFuck out into its own test.
